pyproject.toml
[project]
name = "museum-book-prototype"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "Artur Borecki", email = "me@pufereq.pl" }]
requires-python = ">=3.12"
dependencies = [
    "moviepy>=2.2.1",
    "pygame-ce>=2.5.6",
    "pyserial>=3.5",
    "pyyaml>=6.0.3",
]

[build-system]
requires = ["uv_build>=0.9.10,<0.10.0"]
build-backend = "uv_build"
